rename energy_branches -> supply_branches
